requirements.txt
# Trading Bot Dependencies

# IBKR API
ib-insync==0.9.86

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Configuration & validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0.1

# Market calendars
pandas-market-calendars>=4.3.0

# Scheduling
APScheduler>=3.10.4

# Database
SQLAlchemy>=2.0.23
sqlmodel>=0.0.14

# Logging
structlog>=23.2.0
colorama>=0.4.6

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

